requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.1
async-timeout==4.0.2
asyncer==0.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
cycler==0.11.0
fastapi==0.72.0
filelock==3.7.0
filetype==1.0.9
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
gdown==4.4.0
h11==0.13.0
idna==3.3
imageio==2.19.1
kiwisolver==1.4.2
llvmlite==0.38.0
matplotlib==3.5.2
multidict==6.0.2
networkx==2.8
numba==0.55.1
numpy==1.21.5
onnxruntime==1.10.0
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
protobuf==3.20.1
pydantic==1.9.0
PyMatting==1.1.5
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
PyWavelets==1.3.0
rembg==2.0.13
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.17.1
tifffile==2022.5.4
tk==0.1.0
tkinterdnd2==0.3.0
tqdm==4.62.3
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.0
watchdog==2.1.7
yarl==1.7.2
